basic refactoring done
